feat: Refactor code to merge remote devices and users
This commit refactors the receiveGossip function in the routes.go file to improve the process of merging remote devices and users. It updates the logic to handle merging ESP devices and users based on their updated timestamps. The code now correctly updates the local device's secret key and last request time if the remote device has a more recent update. Similarly, it updates the local user's password hash if the remote user has a more recent update. These changes enhance the functionality of the server by ensuring that the most up-to-date information is stored for each device and user.